redeveloped structure
we will need dadta from both the recruiting and rosters endpoints. the rosters endpoint does not contain data on the players star ratings, so we'll have to collect that from the recruiting endpoint and merge them on name later on. this commit works to get just the relevant data from the recruiting api