integrated get_roster function
added in our get roster function so we are now collecting thhe data for all recruits for certain teams over a 6 year time span, which should cover all players. we are now gonna need to merge the two dataframes we have here. i found a recruiting_id in the rosters endpoint in the cfbd documentation, but from my testing here it seems incomplete. I think im gonna try and merge on names to begin and see where that takes us